frontend: Convert SurfaceAreaPriceCeilingTable to a common component
