+++ frontend/src/styles/base/_general.sass
@@ -143,3 +143,23 @@ ul, ol
   @include flex-flow(row nowrap)
   @include align-items(center)
   gap: $spacing-m
+
+.surface-area-price-ceiling-container
+  @include flexbox()
+  @include flex-flow(row nowrap)
+  width: 100%
+  padding: 0
+  gap: 2rem
+
+  .filter-container
+    width: 20%
+    .help-text
+      color: $color-black-60
+      font-size: $fontsize-body-s
+      line-height: $lineheight-m
+
+  [class^="Table-module_container"]
+    width: 80%
+
+  [class*="LoadingSpinner-module_loadingSpinner"]
+    --spinner-color: var(--color-engel)
