+++ frontend/src/styles/components/_Reports.sass
@@ -48,29 +48,9 @@
       margin-top: 1rem
       margin-bottom: 0.5rem
 
-    &.surface-area-price-ceiling-report-container
+    &.surface-area-price-ceiling-report
       gap: 1rem
       width: 100%
-
-      .surface-area-price-ceiling-list
-        @include flexbox()
-        @include flex-flow(row nowrap)
-        width: 100%
-        padding: 0
-        gap: 2rem
-
-        .filter-container
-          width: 20%
-          .help-text
-            color: $color-black-60
-            font-size: $fontsize-body-s
-            line-height: $lineheight-m
-
-        [class^="Table-module_container"]
-          width: 80%
-
-        [class*="LoadingSpinner-module_loadingSpinner"]
-          --spinner-color: var(--color-engel)
 
     .date-input-wrap
       @include flexbox()
